first

+++ backend/controllers/goalController.js
@@ -7,6 +7,7 @@ const User = require("../models/userModel");
 //Get Goals
 // @route GET/api/goals
 //@access Private
+
 const getGoals = asyncHandler(async (req, res) => {
   const goals = await Goal.find({user:req.user.id});
 
@@ -16,6 +17,7 @@ const getGoals = asyncHandler(async (req, res) => {
 //Set Goals
 // @route POST/api/goals
 //@access Private
+
 const setGoal = asyncHandler(async (req, res) => {
   if (!req.body.text) {
     res.status(400);
@@ -27,6 +29,7 @@ const setGoal = asyncHandler(async (req, res) => {
   });
   res.status(200).json(goal);
 });
+
 //PUT Goals
 // @route PUT/api/goals/:id
 //@access Private
@@ -38,15 +41,14 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
   
-  const user = await User.findById(req.user.id)
 
 //check for user
-  if(!user){
+  if(!req.user){
     res.status(401)
     throw new Error('User Not Found')
   }
   //make sure the logged in user matches the goal user
-  if(goal.user.toString()!== user.id){
+  if(goal.user.toString()!== req.user.id){
     res.status(401)
     throw new Error('User not authorized')
   }
@@ -68,6 +70,11 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+  if(goal.user.toString() !== req.user.id){
+    res.status(401)
+    throw new Error('User not authorized!')
+  }
+
   await goal.deleteOne({ _id: req.params.id });
 
   res.status(200).json({ id: req.params.id });
